requirements.txt
torch>=1.6.0
torchvision
opencv-python
pillow
scikit-image
tensorboard
tqdm